Split SQL AST definitions
